+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingForeignKeyConditionModel.java
@@ -20,7 +20,6 @@ package is.codion.swing.framework.model;
 
 import is.codion.common.Operator;
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.observer.Mutable;
 import is.codion.common.observer.Observer;
 import is.codion.common.state.State;
 import is.codion.common.value.Value;
@@ -100,7 +99,7 @@ public final class SwingForeignKeyConditionModel implements ConditionModel<Entit
 	}
 
 	@Override
-	public Mutable<Wildcard> wildcard() {
+	public Value<Wildcard> wildcard() {
 		return condition.wildcard();
 	}